+++ src/movie-adapter.ts
@@ -10,8 +10,6 @@ import type {
   UpdateMovieResponse
 } from './@types'
 import type { MovieRepository } from './movie-repository'
-import { CreateMovieSchema, UpdateMovieSchema } from './@types/schema'
-import type { ZodSchema } from 'zod'
 
 // ports & adapters (hexagonal) pattern refactor:
 // movies.ts (now called movie-service) has been split into two parts,
@@ -186,14 +184,6 @@ export class MovieAdapter implements MovieRepository {
     id?: number
   ): Promise<CreateMovieResponse | ConflictMovieResponse> {
     try {
-      // Zod Key feature 3: safeParse
-      // Zod note: if you have a frontend, you can use the schema + safeParse there
-      // in order to perform form validation before sending the data to the server
-      const validationResult = validateSchema(CreateMovieSchema, data)
-      if (!validationResult.success) {
-        return { status: 400, error: validationResult.error }
-      }
-
       // Check if the movie already exists
       const existingMovie = await this.prisma.movie.findFirst({
         where: { name: data.name }
@@ -225,13 +215,6 @@ export class MovieAdapter implements MovieRepository {
     UpdateMovieResponse | MovieNotFoundResponse | ConflictMovieResponse
   > {
     try {
-      // Zod Key feature 3: safeParse
-      // Zod note: if you have a frontend, you can use the schema + safeParse there
-      // in order to perform form validation before sending the data to the server
-      const validationResult = validateSchema(UpdateMovieSchema, data)
-      if (!validationResult.success)
-        return { status: 400, error: validationResult.error }
-
       const existingMovie = await this.prisma.movie.findUnique({
         where: { id }
       })
@@ -251,21 +234,5 @@ export class MovieAdapter implements MovieRepository {
       this.handleError(error)
       return { status: 500, error: 'Internal server error' }
     }
-  }
-}
-
-// helper function for schema validation
-function validateSchema<T>(
-  schema: ZodSchema<T>,
-  data: unknown
-): { success: true; data: T } | { success: false; error: string } {
-  const result = schema.safeParse(data)
-  if (result.success) {
-    return { success: true, data: result.data }
-  } else {
-    const errorMessages = result.error.errors
-      .map((err) => err.message)
-      .join(', ')
-    return { success: false, error: errorMessages }
   }
 }
